API-ML/requirements.txt
fastapi
pydantic
uvicorn[standard]
python-multipart
numpy
Pillow
tensorflow
requests
fastapi.security
